BlogPostServiceTest & ImageFileStorageServiceTest - moved under service folder
