+++ src/test/java/com/wenglam/baking_app/service/BlogPostServiceTest.java
@@ -1,4 +1,4 @@
-package com.wenglam.baking_app;
+package com.wenglam.baking_app.service;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
